Move Todolist components each components folder
