+++ client/app/(main)/todolist/components/DraggableTodolist.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { DndContext } from '@dnd-kit/core'
 import { SortableContext } from '@dnd-kit/sortable'
-import { SortableOverlay, TodoItem } from '@/app/ui'
+import { SortableOverlay } from '@/app/ui'
+import { TodoItem } from '@/app/(main)/todolist/components'
 import { saveTodolistOrder, useDragDndKit } from '@/app/utils'
 import { Todo } from '@/app/types'
